🚚 Split up declarations
