+++ Packages/Core/Sources/API/Common/Query/QueryComparable.swift
@@ -1,0 +1,15 @@
+//
+//  QueryComparable.swift
+//  Core
+//
+//  Created by Dorian on 13/05/2025.
+//
+
+import Foundation
+
+public protocol QueryComparable: QueryFilterable, Comparable {}
+
+extension String: QueryComparable {}
+extension Int: QueryComparable {}
+extension Double: QueryComparable {}
+extension Date: QueryComparable {}
