+++ Packages/Core/Sources/API/Common/Query/QueryFilterable.swift
@@ -1,0 +1,16 @@
+//
+//  QueryFilterable.swift
+//  Core
+//
+//  Created by Dorian on 13/05/2025.
+//
+
+import Foundation
+
+public protocol QueryFilterable: Sendable, Equatable, Codable {}
+
+extension String: QueryFilterable {}
+extension Int: QueryFilterable {}
+extension Double: QueryFilterable {}
+extension Date: QueryFilterable {}
+extension Bool: QueryFilterable {}
